[Refactor] 이미지 아이콘에 memo 적용
