+++ src/components/index.js
@@ -1,5 +1,7 @@
 import MonthCalendar from "./MonthCalendar";
 import ImageButton from "./ImageButton";
 import PinkButton from "./PinkButton";
+import EmotionRatio from './EmotionRatio';
+import LocationRank from './LocationRank';
 
-export {MonthCalendar, ImageButton, PinkButton};
+export {MonthCalendar, ImageButton, PinkButton, EmotionRatio, LocationRank};
